feat(vehicles): added vehicle component
