+++ app/src/views/User/Vehicles/Vehicles.scss
@@ -1,0 +1,187 @@
+@import "../../../styles/colors.scss";
+
+.cars-container {
+  display: grid;
+  width: 100%;
+  gap: 20px;
+
+  .ant-pagination-item-link:hover,
+  .ant-pagination-item:hover,
+  .ant-pagination-item a:hover,
+  .ant-pagination-item-active,
+  .ant-pagination-item-active a {
+    color: $secondary-color;
+    border-color: $secondary-color;
+  }
+
+  .ant-pagination-item-link:hover,
+  .ant-pagination-item:hover,
+  .ant-pagination-item a:hover,
+  .ant-pagination-item-active,
+  .ant-pagination-item-active a {
+    color: $secondary-color;
+    border-color: $secondary-color;
+  }
+
+  .container-listing .ant-select-selector,
+  .pagination-container .ant-select-selector {
+    border: none;
+    box-shadow: none !important;
+  }
+
+  .pagination-container {
+    display: flex;
+    align-items: center;
+    justify-content: space-between;
+
+    span {
+      margin-bottom: 2px;
+    }
+  }
+
+  .search-container {
+    display: flex;
+    justify-content: center;
+
+    .search-input {
+      width: 40%;
+      min-width: 200px;
+
+      input {
+        height: 45px;
+        border-radius: 5px !important;
+
+        &:focus {
+          box-shadow: none;
+        }
+
+        &::placeholder {
+          color: $grey-900;
+        }
+      }
+
+      span.ant-input-wrapper.ant-input-group {
+        display: flex;
+        flex-direction: row-reverse;
+        align-items: center;
+      }
+
+      .search-input .ant-input-group .ant-input {
+        z-index: 0;
+        width: 100%;
+        margin-bottom: 0;
+        padding-left: 40px;
+        background: $grey-300 0% 0% no-repeat padding-box !important;
+        border: none !important;
+      }
+
+      button.ant-btn.ant-btn-icon-only.ant-input-search-button {
+        background: transparent;
+        border: none;
+      }
+
+      span.anticon.anticon-search {
+        color: $grey-900;
+        padding-left: 10px;
+      }
+    }
+  }
+
+  .viewer-container {
+    width: 100%;
+    display: flex;
+    align-items: center;
+    justify-content: center;
+  }
+
+  .container-listing {
+    width: 85%;
+    display: grid;
+    justify-content: center;
+    grid-template-columns: 25% 25% 25% 25%;
+    gap: 4rem;
+
+    .car-container {
+      color: $text-color;
+      background-color: $grey-300;
+      padding: 25px;
+      display: grid;
+      gap: 10px;
+      text-align: center;
+    
+      .image-container {
+        display: flex;
+        align-items: center;
+        justify-content: center;
+      }
+    
+      .image-container > img {
+        width: 80%;
+      }
+    }
+    
+    .offer-button {
+      display: flex;
+      align-items: center;
+      justify-content: center;
+      font-weight: bold;
+      border: none;
+      font-size: 24px;
+      padding: 25px 30px;
+      color: $text-color;
+      background-color: $primary-color;
+      transition: all 300ms ease-in-out;
+    
+      &:hover #value {
+        display: inline;
+      }
+    
+      &:hover #text {
+        display: none;
+      }
+    
+      #value {
+        display: none;
+        font-size: 24px;
+        font-weight: bold;
+        margin-right: 12px;
+      }   
+    }
+    
+    .offer-button:focus,
+    .offer-button:hover {
+      border: none;
+      color: $text-color;
+      background-color: $hover-primary;
+    }
+  }
+}
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
